model/requirements.txt
pillow
google.generativeai
gtts
deep_translator
python-dotenv
flask
langdetect
flask-cors